Rename select helper to match convention
