+++ lib/govuk_publishing_components.rb
@@ -21,7 +21,7 @@ require "govuk_publishing_components/presenters/content_breadcrumbs_based_on_pri
 require "govuk_publishing_components/presenters/content_breadcrumbs_based_on_taxons"
 require "govuk_publishing_components/presenters/content_breadcrumbs_based_on_topic"
 require "govuk_publishing_components/presenters/checkboxes_helper"
-require "govuk_publishing_components/presenters/select"
+require "govuk_publishing_components/presenters/select_helper"
 require "govuk_publishing_components/presenters/meta_tags"
 require "govuk_publishing_components/presenters/content_item"
 require "govuk_publishing_components/presenters/translation_nav_helper"
